requirements.txt
requests==2.6.0
xmltodict==0.9.2
xmldict==0.4.1
python-sshtail==0.0.2
Fabric==1.8.3
bumpversion==0.5.1
setuptools==12.0.5
